// gauge-ts/src/gen/messages.ts
export interface StepValidateRequest {
  /** / The text is used to lookup Step implementation */
  stepText: string;
  /** / The number of paramters in the Step */
  numberOfParameters: number;
  /** /This is use to generate step implementation template */
  stepValue: ProtoStepValue | undefined;
}

/**
 * / Response of StepValidateRequest.
 * / The runner tells the caller if the Request was valid,
 * / i.e. an implementation exists for given Step text.
 * / Returns an error message if it is an error response.
 */
export interface StepValidateResponse {
  isValid: boolean;
  errorMessage: string;
  errorType: StepValidateResponse_ErrorType;
  suggestion: string;
}

export enum StepValidateResponse_ErrorType {
  STEP_IMPLEMENTATION_NOT_FOUND = 0,
  DUPLICATE_STEP_IMPLEMENTATION = 1,
  UNRECOGNIZED = -1,
}
// ...
export const StepValidateRequest: MessageFns<StepValidateRequest> = {
  encode(message: StepValidateRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.stepText !== "") {
      writer.uint32(10).string(message.stepText);
    }
    if (message.numberOfParameters !== 0) {
      writer.uint32(16).int32(message.numberOfParameters);
    }
    if (message.stepValue !== undefined) {
      ProtoStepValue.encode(message.stepValue, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): StepValidateRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseStepValidateRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.stepText = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.numberOfParameters = reader.int32();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.stepValue = ProtoStepValue.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): StepValidateRequest {
    return {
      stepText: isSet(object.stepText) ? globalThis.String(object.stepText) : "",
      numberOfParameters: isSet(object.numberOfParameters) ? globalThis.Number(object.numberOfParameters) : 0,
      stepValue: isSet(object.stepValue) ? ProtoStepValue.fromJSON(object.stepValue) : undefined,
    };
  },

  toJSON(message: StepValidateRequest): unknown {
    const obj: any = {};
    if (message.stepText !== "") {
      obj.stepText = message.stepText;
    }
    if (message.numberOfParameters !== 0) {
      obj.numberOfParameters = Math.round(message.numberOfParameters);
    }
    if (message.stepValue !== undefined) {
      obj.stepValue = ProtoStepValue.toJSON(message.stepValue);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<StepValidateRequest>, I>>(base?: I): StepValidateRequest {
    return StepValidateRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<StepValidateRequest>, I>>(object: I): StepValidateRequest {
    const message = createBaseStepValidateRequest();
    message.stepText = object.stepText ?? "";
    message.numberOfParameters = object.numberOfParameters ?? 0;
    message.stepValue = (object.stepValue !== undefined && object.stepValue !== null)
      ? ProtoStepValue.fromPartial(object.stepValue)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}